+++ main.tf
@@ -1,24 +1,34 @@
-
 data "template_file" "nomad_job_redash_server" {
   template = file("${path.module}/conf/nomad/redash_server.hcl")
-
   vars = {
+    postgres_service_name = var.postgres_service_name
+    postgres_port         = var.postgres_port
+    postgres_user         = var.postgres_username
+    postgres_password     = var.postgres_password
     redash_admin_username = var.redash_admin_username
     redash_admin_password = var.redash_admin_password
-    redash_admin_email_id  = var.redash_admin_email_id
+    redash_admin_email_id = var.redash_admin_email_id
   }
 }
 
 data "template_file" "nomad_job_redash_worker" {
   template = file("${path.module}/conf/nomad/redash_worker.hcl")
+  vars = {
+    postgres_service_name = var.postgres_service_name
+    postgres_port         = var.postgres_port
+    postgres_user         = var.postgres_username
+    postgres_password     = var.postgres_password
+  }
 }
 
 data "template_file" "nomad_job_redash_scheduler" {
   template = file("${path.module}/conf/nomad/redash_scheduler.hcl")
-}
-
-data "template_file" "nomad_job_postgres" {
-  template = file("${path.module}/conf/nomad/postgres.hcl")
+  vars = {
+    postgres_service_name = var.postgres_service_name
+    postgres_port         = var.postgres_port
+    postgres_user         = var.postgres_username
+    postgres_password     = var.postgres_password
+  }
 }
 
 data "template_file" "nomad_job_redis" {
@@ -34,7 +44,6 @@ resource "nomad_job" "nomad_job_redash_server" {
   detach = false
   depends_on = [
     nomad_job.nomad_job_redis,
-    nomad_job.nomad_job_postgres,
     nomad_job.nomad_job_email
   ]
 }
@@ -55,11 +64,6 @@ resource "nomad_job" "nomad_job_redash_scheduler" {
   ]
 }
 
-resource "nomad_job" "nomad_job_postgres" {
-  jobspec = data.template_file.nomad_job_postgres.rendered
-  detach = false
-}
-
 resource "nomad_job" "nomad_job_redis" {
   jobspec = data.template_file.nomad_job_redis.rendered
   detach = false
@@ -68,5 +72,4 @@ resource "nomad_job" "nomad_job_redis" {
 resource "nomad_job" "nomad_job_email" {
   jobspec = data.template_file.nomad_job_email.rendered
   detach = false
-
 }